# api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Tapis Notifications API
  description: Tapis Notifications API definition
  contact:
    name: CicSupport
    url: http://tacc.utexas.edu
    email: cicsupport@tacc.utexas.edu
  license:
    name: Apache 2.0
    url: http://foo.bar
  version: "1.0"
servers:
- url: http://localhost:8080/
  description: localhost
  variables: {}
- url: https://dev.develop.tapis.io
  description: development
  variables: {}
security:
- x-tapis-token: []
...
paths:
...
  /v3/notifications/queues/{queueId}:
    delete:
      tags:
      - queues
      summary: Delete a queue.
      operationId: deleteTopic
      parameters:
      - name: queueId
        in: path
        description: ID of the queue
        required: true
        schema:
          type: string
        example: queuue12345
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QueueResponse'
        401:
          description: Not Authenticated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QueueResponse'
        403:
          description: Not Authorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QueueResponse'
        500:
          description: Internal Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QueueResponse'
  /v3/notifications/topics/{topicName}:
    post:
      tags:
      - topics
      summary: Create a new Notification in the topic channel
      operationId: sendNotification
      parameters:
      - name: topicName
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
    delete:
      tags:
      - topics
      summary: Delete a topic. Note, this will also delete any subscriptions that
        are attached to the topic
      operationId: deleteTopic_1
      parameters:
      - name: topicName
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
  /v3/notifications/topics/{topicId}:
    get:
      tags:
      - topics
      summary: Get details of a topic.
      operationId: getTopic
      parameters:
      - name: topicId
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
  /v3/notifications/topics:
    get:
      tags:
      - topics
      summary: Get a list of all topics available to you.
      operationId: getTopics
      parameters:
      - name: limit
        in: query
        description: pagination limit
        schema:
          type: integer
          format: int32
          default: 1000
        example: 100
      - name: offset
        in: query
        description: pagination offset
        schema:
          type: integer
          format: int64
          default: 0
        example: 1000
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
    post:
      tags:
      - topics
      summary: Create a new topic to publish message to
      operationId: createTopic
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateTopicRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
  /v3/notifications/topics/{topicId}/subscriptions:
    get:
      tags:
      - subscriptions
      summary: Get a list of all subscriptions on this topic
      operationId: getSubscriptions
      parameters:
      - name: topicId
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
    post:
      tags:
      - subscriptions
      summary: Create a new subscription to this topic channel
      operationId: createSubscription
      parameters:
      - name: topicId
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
  /v3/notifications/topics/{topicId}/subscriptions/{subscriptionId}:
    get:
      tags:
      - subscriptions
      summary: Get a subscription by ID
      operationId: getSubscriptionByID
      parameters:
      - name: topicId
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      - name: subscriptionId
        in: path
        description: ID of the subscription
        required: true
        schema:
          type: string
        example: 1234-123-123
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
    delete:
      tags:
      - subscriptions
      summary: Delete a subscription
      operationId: deleteSubscription
      parameters:
      - name: topicId
        in: path
        description: ID of the topic
        required: true
        schema:
          type: string
        example: mySuperTopic
      - name: subscriptionId
        in: path
        description: ID of the subscription
        required: true
        schema:
          type: string
        example: 1234-123-123
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
...
components:
  schemas:
    QueueResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
        version:
          type: string
    Topic:
      type: object
      properties:
        id:
          type: integer
          format: int32
        tenantId:
          type: string
        uuid:
          type: string
          format: uuid
        name:
          type: string
        created:
          type: string
          format: date-time
        owner:
          type: string
        description:
          type: string
        schema:
          type: object
          additionalProperties:
            type: object
            description: A valid JSON schema
          description: A valid JSON schema
...
    CreateTopicRequest:
      required:
      - description
      - name
      - schema
      type: object
      properties:
        name:
          type: string
          description: Name for the topic, must be unique
        schema:
          type: string
          description: A valid JSON schema
        description:
          type: string
          description: An informative description
...
    CreateNotificationRequest:
      required:
      - body
      - id
      type: object
      properties:
        body:
          type: string
        id:
          type: string
    Notification:
      required:
      - data
      - source
      - specversion
      - subject
      - tenantId
      - time
      - type
      - uuid
      type: object
      properties:
        specversion:
          type: string
        uuid:
          type: string
          format: uuid
        tenantId:
          type: string
        time:
          type: integer
          format: int64
        source:
          type: string
        data:
          type: object
        subject:
          type: string
        type:
          type: string
    NotificationResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          $ref: '#/components/schemas/Notification'
        version:
          type: string
    TopicResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          $ref: '#/components/schemas/Topic'
        version:
          type: string
    TopicListResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
        version:
          type: string
    TapisResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: object
        version:
          type: string
...
    SubscriptionListResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        version:
          type: string
    CreateSubscriptionRequest:
      required:
      - notificationMechanisms
      type: object
      properties:
        filter:
          type: string
        notificationMechanisms:
          type: array
          items:
            $ref: '#/components/schemas/NotificationMechanism'